archive/requirements.txt
# ray
ray[data,train,tune,serve]==2.48
pynvml

# torch
torch
torchvision
torchmetrics

# HF
huggingface_hub[cli]

# vllm
vllm
#flashinfer-python

# jupyter
nbformat
ipykernel
ipywidgets

# upnp
miniupnpc

# fastapi
fastapi[standard]
pydantic_settings
sqlmodel

# wrapt
wrapt

# redis
redis

# misc
tqdm
click
pandas
pyarrow
openpyxl
colorama
matplotlib
